+++ packages/odyssey-react-mui/src/MenuButton.tsx
@@ -41,6 +41,10 @@ export interface MenuButtonProps {
    * The variant of the triggering Button
    */
   buttonVariant?: ButtonProps["variant"];
+  /**
+   * The id of the `input` element.
+   */
+  id?: string;
 }
 
 const MenuButton = ({
@@ -48,6 +52,7 @@ const MenuButton = ({
   children,
   buttonEndIcon = <ChevronDownIcon />,
   buttonVariant = "secondary",
+  id: idOverride,
 }: MenuButtonProps) => {
   const [anchorEl, setAnchorEl] = useState<null | HTMLElement>(null);
 
@@ -61,7 +66,7 @@ const MenuButton = ({
     setAnchorEl(null);
   };
 
-  const uniqueId = useUniqueId();
+  const uniqueId = useUniqueId(idOverride);
 
   const menuListProps = useMemo(
     () => ({ "aria-labelledby": `${uniqueId}-button` }),
@@ -77,10 +82,9 @@ const MenuButton = ({
         aria-haspopup="true"
         aria-expanded={open ? "true" : undefined}
         onClick={handleClick}
+        text={buttonLabel}
         variant={buttonVariant}
-      >
-        {buttonLabel}
-      </Button>
+      />
       <Menu
         id={`${uniqueId}-menu`}
         anchorEl={anchorEl}